backend/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ─── HTTP client ──────────────────────────────────────────────────────────────
httpx==0.27.2

# ─── Auth0 / JWT ──────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0

# ─── Config / env ─────────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
